make check & verify routes more robust
Finally, `check.js` really checks if there has been a request with the
given e-mail and address. This should prevent mass retrieval at least
in parts. Moving on, we might put a time limit for this in place.

`verify.js` now only stores the e-mail in the DB if sending the puzzle
has been successful. Therefore, we can assume that the user has actually
received the e-mail if the entry is in the DB. This should reduce the
number of "You already requested verification" errors in the UI, which
are caused by aborted e-mail verification processes (and the UI not
being able to handle them properly).